Speedtest between Java and Python for generating prime numbers.

+++ src/java/richardbergquist/helloworld/prime/Prime.java
@@ -16,11 +16,11 @@ public class Prime {
   public final static void main(String args[]) {
 
     //Parse commandline args
-    int startFromValue = 0;
-    int delta = 0;
+    long startFromValue = 0;
+    long delta = 0;
     try {
-      startFromValue = Integer.parseInt(args[0]);
-      delta = Integer.parseInt(args[1]);
+      startFromValue = Long.parseLong(args[0]);
+      delta = Long.parseLong(args[1]);
     } catch (NumberFormatException exNumberFormat) {
       System.out.println(exNumberFormat.toString());
       System.exit(-1);
@@ -39,17 +39,17 @@ public class Prime {
    * int startFromValue : The value to to start calculating primes from
    * int delta: The value when added to startFromValue, to stop calculating primes.
    */
-  private void printprimes(int startFromValue, int delta) {
+  private void printprimes(long startFromValue, long delta) {
 
     System.out.println("Generating Primes");
     System.out.println("startFromValue: " + startFromValue);
     System.out.println("delta: " + delta);
 
-    int limit = startFromValue + delta;
-    for (int n = startFromValue ; n <= limit; n++) {
+    long limit = startFromValue + delta;
+    for (long n = startFromValue ; n <= limit; n++) {
       boolean founddivisor = false;
 
-      for (int x = 2 ; x < n ; x ++) {
+      for (long x = 2 ; x < n ; x ++) {
         if ( n % x == 0 ) {
           // found a divisor
           founddivisor = true;
